First version and documentation.
This is a fork of the popular MSYS2 sshd setup script gist:

https://gist.github.com/samhocevar/00eec26d9e9988d080ac

The script was modified to install the necessary packages automatically,
generate the host keys, edit the sshd_config to set the port to 2222 and
add a firewall rule.

The README.md describes the necessary setup to use this from ssh and
terminal preview.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>