More automation, Cygwin support.
Add support for Cygwin, using mostly the same process, installing
dependencies via a slightly modified apt-cyg.

Both the Cygwin and the MSYS2 install work fine on unmodified default
installs.

By default, the port for MSYS2 is 2222 and 2223 for Cygwin, both the
MSYS2 and the Cygwin sshds can be used simultaneously, as all the
necessary variables like users and service names are distinct.

Add a --port or -p option to specify the port.

Generate an ssh key if the user does not have one, and automatically add
the user's keys to authorized_keys.

Automatically add aliases to ~/.ssh/config.

Make sure the last line of ~/.bash_logout is `exit 0` to prevent the
terminal tab from not closing because of a non-zero exit status from
ssh.

- Fix #1

Signed-off-by: Rafael Kitover <rkitover@gmail.com>