Finish domain support, home dir info, errs to >&2.
The script is now tested to work on computers joined to a domain, add
the only missing piece which is to run passwd -R, as suggested by the
original author.

Send all errors and warnings to STDERR.

Add a doc section about how to use the Windows profile directory as the
MSYS2/Cygwin home directory.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>