More domain fixes, symlinks support #2.
Append mkpasswd -c (current user) to /etc/passwd, this may be necessary
for domain users which will not be listed in mkpasswd -l (local users.)

Likewise, append mkgroup -c to /etc/group and remove duplicates.

Set the runtime environment variable, which is CYGWIN for Cygwin and
MSYS for MSYS2, on service startup with "ntsec" which may be necessary
for domain user logins.

Also add "export" and "wincmdln" to the runtime env var, these are nice
to have.

If the system is in developer mode, also add "winsymlinks:nativestrict"
to enable real symlink support.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>