Fix download instructions, suppress dev mode err.
Provide separate script download instructions for Powershell and
bash/cmd.exe, as in Windows Powershell, as opposed to Powershell Core,
curl is an alias for invoke-webrequest and not the windows curl program.

Suppress error messages from the reg.exe invocation that checks for
developer mode, and add a message that symlinks are being enabled if
developer mode is enabled.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>